EPMCUXDA-6612: Moved euring related entities to a separate folder
