+++ src/entities/euring-codes/accuracy-of-coords-intity.ts
@@ -1,11 +1,11 @@
 import { Entity, PrimaryColumn, Column, OneToMany } from 'typeorm';
 import { Dictionary } from './common-interfaces';
-import { Ring } from './ring-entity';
+import { Ring } from '../ring-entity';
 
 @Entity()
-export class EURINGOtherMarksInformation implements Dictionary {
+export class AccuracyOfCoords implements Dictionary {
   @PrimaryColumn()
-  public id: string;
+  public id: number;
 
   @Column('varchar', { nullable: true, default: null })
   public desc_eng: string | null;
@@ -16,6 +16,6 @@ export class EURINGOtherMarksInformation implements Dictionary {
   @Column('varchar', { nullable: true, default: null })
   public desc_byn: string | null;
 
-  @OneToMany(() => Ring, m => m.euringOtherMarksInformation)
+  @OneToMany(() => Ring, m => m.accuracyOfCoords)
   public ring: Ring[];
 }
