+++ src/entities/euring-codes/accuracy-of-date-entity.ts
@@ -1,11 +1,11 @@
 import { Entity, PrimaryColumn, Column, OneToMany } from 'typeorm';
 import { Dictionary } from './common-interfaces';
-import { Ring } from './ring-entity';
+import { Ring } from '../ring-entity';
 
 @Entity()
-export class CathingLures implements Dictionary {
+export class AccuracyOfDate implements Dictionary {
   @PrimaryColumn()
-  public id: string;
+  public id: number;
 
   @Column('varchar', { nullable: true, default: null })
   public desc_eng: string | null;
@@ -16,6 +16,6 @@ export class CathingLures implements Dictionary {
   @Column('varchar', { nullable: true, default: null })
   public desc_byn: string | null;
 
-  @OneToMany(() => Ring, m => m.cathingLures)
+  @OneToMany(() => Ring, m => m.accuracyOfDate)
   public ring: Ring[];
 }
