+++ src/entities/euring-codes/age-entity.ts
@@ -1,11 +1,13 @@
 import { Entity, PrimaryColumn, Column, OneToMany } from 'typeorm';
 import { Dictionary } from './common-interfaces';
-import { Ring } from './ring-entity';
+import { Ring } from '../ring-entity';
+import { BasaRing } from '../basa-ring-entity';
+import { Observation } from '../observation-entity';
 
 @Entity()
-export class VerificationOfTheMetalRing implements Dictionary {
+export class Age implements Dictionary {
   @PrimaryColumn()
-  public id: number;
+  public id: string;
 
   @Column('varchar', { nullable: true, default: null })
   public desc_eng: string | null;
@@ -16,6 +18,12 @@ export class VerificationOfTheMetalRing implements Dictionary {
   @Column('varchar', { nullable: true, default: null })
   public desc_byn: string | null;
 
-  @OneToMany(() => Ring, m => m.verificationOfTheMetalRing)
+  @OneToMany(() => Ring, m => m.ageScheme)
   public ring: Ring[];
+
+  @OneToMany(() => Observation, m => m.speciesMentioned)
+  public observation: Observation[];
+
+  @OneToMany(() => BasaRing, m => m.age)
+  public basaRing: BasaRing[];
 }
