+++ src/entities/euring-codes/broodsize-entity.ts
@@ -1,11 +1,11 @@
 import { Entity, PrimaryColumn, Column, OneToMany } from 'typeorm';
 import { Dictionary } from './common-interfaces';
-import { Ring } from './ring-entity';
+import { Ring } from '../ring-entity';
 
 @Entity()
-export class Manipulated implements Dictionary {
+export class Broodsize implements Dictionary {
   @PrimaryColumn()
-  public id: string;
+  public id: number;
 
   @Column('varchar', { nullable: true, default: null })
   public desc_eng: string | null;
@@ -16,6 +16,6 @@ export class Manipulated implements Dictionary {
   @Column('varchar', { nullable: true, default: null })
   public desc_byn: string | null;
 
-  @OneToMany(() => Ring, m => m.manipulated)
+  @OneToMany(() => Ring, m => m.broodsize)
   public ring: Ring[];
 }
