+++ src/entities/euring-codes/circumstances-entity.ts
@@ -1,9 +1,9 @@
 import { Entity, PrimaryColumn, Column, OneToMany } from 'typeorm';
 import { Dictionary } from './common-interfaces';
-import { Ring } from './ring-entity';
+import { Ring } from '../ring-entity';
 
 @Entity()
-export class AccuracyOfDate implements Dictionary {
+export class Circumstances implements Dictionary {
   @PrimaryColumn()
   public id: number;
 
@@ -16,6 +16,6 @@ export class AccuracyOfDate implements Dictionary {
   @Column('varchar', { nullable: true, default: null })
   public desc_byn: string | null;
 
-  @OneToMany(() => Ring, m => m.accuracyOfDate)
+  @OneToMany(() => Ring, m => m.circumstances)
   public ring: Ring[];
 }
