+++ src/entities/euring-codes/color-ring-information-entity.ts
@@ -1,21 +1,20 @@
 import { Entity, PrimaryColumn, Column, OneToMany } from 'typeorm';
-import { Dictionary } from './common-interfaces';
-import { Ring } from './ring-entity';
+import { Ring } from '../ring-entity';
 
 @Entity()
-export class EURINGPrimaryIdentificationMethod implements Dictionary {
+export class ColorRingInformation {
   @PrimaryColumn()
   public id: string;
 
   @Column('varchar', { nullable: true, default: null })
-  public desc_eng: string | null;
+  public schem: string | null;
 
   @Column('varchar', { nullable: true, default: null })
-  public desc_rus: string | null;
+  public descr: string | null;
 
   @Column('varchar', { nullable: true, default: null })
-  public desc_byn: string | null;
+  public pic: string | null;
 
-  @OneToMany(() => Ring, m => m.euringPrimaryIdMethod)
+  @OneToMany(() => Ring, m => m.colorRingInformation)
   public ring: Ring[];
 }
