+++ src/entities/euring-codes/common-interfaces.ts
