+++ src/entities/euring-codes/euring-other-marks-information-entity.ts
@@ -1,9 +1,9 @@
 import { Entity, PrimaryColumn, Column, OneToMany } from 'typeorm';
 import { Dictionary } from './common-interfaces';
-import { Ring } from './ring-entity';
+import { Ring } from '../ring-entity';
 
 @Entity()
-export class CatchingMethod implements Dictionary {
+export class EURINGOtherMarksInformation implements Dictionary {
   @PrimaryColumn()
   public id: string;
 
@@ -16,6 +16,6 @@ export class CatchingMethod implements Dictionary {
   @Column('varchar', { nullable: true, default: null })
   public desc_byn: string | null;
 
-  @OneToMany(() => Ring, m => m.catchingMethod)
+  @OneToMany(() => Ring, m => m.euringOtherMarksInformation)
   public ring: Ring[];
 }
