+++ src/entities/euring-codes/metal-ring-information-entity.ts
@@ -1,9 +1,9 @@
 import { Entity, PrimaryColumn, Column, OneToMany } from 'typeorm';
 import { Dictionary } from './common-interfaces';
-import { Ring } from './ring-entity';
+import { Ring } from '../ring-entity';
 
 @Entity()
-export class EURINGCodeIdentifier implements Dictionary {
+export class MetalRingInformation implements Dictionary {
   @PrimaryColumn()
   public id: number;
 
@@ -16,6 +16,6 @@ export class EURINGCodeIdentifier implements Dictionary {
   @Column('varchar', { nullable: true, default: null })
   public desc_byn: string | null;
 
-  @OneToMany(() => Ring, m => m.euringCodeIdentifier)
+  @OneToMany(() => Ring, m => m.metalRingInformation)
   public ring: Ring[];
 }
