+++ src/entities/euring-codes/pullus-age-entity.ts
@@ -1,9 +1,9 @@
 import { Entity, PrimaryColumn, Column, OneToMany } from 'typeorm';
 import { Dictionary } from './common-interfaces';
-import { Ring } from './ring-entity';
+import { Ring } from '../ring-entity';
 
 @Entity()
-export class AccuracyOfCoords implements Dictionary {
+export class PullusAge implements Dictionary {
   @PrimaryColumn()
   public id: number;
 
@@ -16,6 +16,6 @@ export class AccuracyOfCoords implements Dictionary {
   @Column('varchar', { nullable: true, default: null })
   public desc_byn: string | null;
 
-  @OneToMany(() => Ring, m => m.accuracyOfCoords)
+  @OneToMany(() => Ring, m => m.pullusAge)
   public ring: Ring[];
 }
