+++ src/entities/euring-codes/ring-scheme-entity.ts
@@ -1,21 +1,20 @@
 import { Entity, PrimaryColumn, Column, OneToMany } from 'typeorm';
-import { Dictionary } from './common-interfaces';
-import { Ring } from './ring-entity';
+import { Ring } from '../ring-entity';
 
 @Entity()
-export class Status implements Dictionary {
+export class RingingScheme {
   @PrimaryColumn()
   public id: string;
 
   @Column('varchar', { nullable: true, default: null })
-  public desc_eng: string | null;
+  public status: string;
 
   @Column('varchar', { nullable: true, default: null })
-  public desc_rus: string | null;
+  public country: string | null;
 
   @Column('varchar', { nullable: true, default: null })
-  public desc_byn: string | null;
+  public center: string | null;
 
-  @OneToMany(() => Ring, m => m.status)
+  @OneToMany(() => Ring, m => m.ringingScheme)
   public ring: Ring[];
 }
