+++ src/entities/euring-codes/sex-entity.ts
@@ -1,11 +1,11 @@
 import { Entity, PrimaryColumn, Column, OneToMany } from 'typeorm';
 import { Dictionary } from './common-interfaces';
-import { Ring } from './ring-entity';
-import { BasaRing } from './basa-ring-entity';
-import { Observation } from './observation-entity';
+import { Ring } from '../ring-entity';
+import { BasaRing } from '../basa-ring-entity';
+import { Observation } from '../observation-entity';
 
 @Entity()
-export class Age implements Dictionary {
+export class Sex implements Dictionary {
   @PrimaryColumn()
   public id: string;
 
@@ -18,12 +18,12 @@ export class Age implements Dictionary {
   @Column('varchar', { nullable: true, default: null })
   public desc_byn: string | null;
 
-  @OneToMany(() => Ring, m => m.ageScheme)
+  @OneToMany(() => Ring, m => m.sexScheme)
   public ring: Ring[];
 
   @OneToMany(() => Observation, m => m.speciesMentioned)
   public observation: Observation[];
 
-  @OneToMany(() => BasaRing, m => m.age)
+  @OneToMany(() => BasaRing, m => m.species)
   public basaRing: BasaRing[];
 }
