+++ src/entities/euring-codes/status-entity.ts
@@ -1,20 +1,21 @@
 import { Entity, PrimaryColumn, Column, OneToMany } from 'typeorm';
-import { Ring } from './ring-entity';
+import { Dictionary } from './common-interfaces';
+import { Ring } from '../ring-entity';
 
 @Entity()
-export class RingingScheme {
+export class Status implements Dictionary {
   @PrimaryColumn()
   public id: string;
 
   @Column('varchar', { nullable: true, default: null })
-  public status: string;
+  public desc_eng: string | null;
 
   @Column('varchar', { nullable: true, default: null })
-  public country: string | null;
+  public desc_rus: string | null;
 
   @Column('varchar', { nullable: true, default: null })
-  public center: string | null;
+  public desc_byn: string | null;
 
-  @OneToMany(() => Ring, m => m.ringingScheme)
+  @OneToMany(() => Ring, m => m.status)
   public ring: Ring[];
 }
