+++ src/entities/observation-entity.ts
@@ -1,9 +1,9 @@
 import { Entity, Column, ManyToOne, PrimaryGeneratedColumn } from 'typeorm';
 import { User } from './user-entity';
 import { Ring } from './ring-entity';
-import { Sex } from './sex-entity';
-import { Age } from './age-entity';
-import { Species } from './species-entity';
+import { Sex } from './euring-codes/sex-entity';
+import { Age } from './euring-codes/age-entity';
+import { Species } from './euring-codes/species-entity';
 
 @Entity()
 export class Observation {
