+++ src/entities/ring-entity.ts
@@ -1,30 +1,30 @@
 import { Entity, PrimaryColumn, Column, ManyToOne, OneToMany } from 'typeorm';
-import { AccuracyOfCoords } from './accuracy-of-coords-intity';
-import { Sex } from './sex-entity';
-import { Age } from './age-entity';
-import { RingingScheme } from './ring-scheme-entity';
-import { EURINGPrimaryIdentificationMethod } from './euring-primary-identification-method-entity';
-import { VerificationOfTheMetalRing } from './verification-of-metal-ring-entity';
-import { MetalRingInformation } from './metal-ring-information-entity';
-import { EURINGOtherMarksInformation } from './euring-other-marks-information-entity';
-import { Species } from './species-entity';
-import { Manipulated } from './manipulated-entity';
-import { MovedBeforeTheCapture } from './moved-before-capture-entity';
-import { CatchingMethod } from './catching-method-entity';
-import { CathingLures } from './catching-lures-entity';
-import { Status } from './status-entity';
-import { Broodsize } from './broodsize-entity';
-import { PullusAge } from './pullus-age-entity';
-import { AccuracyOfPullusAge } from './accuracy-of-pullus-age-entity';
-import { PlaceCode } from './place-code-entity';
-import { Condition } from './condition-entity';
-import { Circumstances } from './circumstances-entity';
-import { CircumstancesPresumed } from './circumstances-presumed-entity';
-import { AccuracyOfDate } from './accuracy-of-date-entity';
-import { EURINGCodeIdentifier } from './euring-code-identifier-entity';
-import { ColorRingInformation } from './color-ring-information-entity';
+import { AccuracyOfCoords } from './euring-codes/accuracy-of-coords-intity';
+import { Sex } from './euring-codes/sex-entity';
+import { Age } from './euring-codes/age-entity';
+import { RingingScheme } from './euring-codes/ring-scheme-entity';
+import { EURINGPrimaryIdentificationMethod } from './euring-codes/euring-primary-identification-method-entity';
+import { VerificationOfTheMetalRing } from './euring-codes/verification-of-metal-ring-entity';
+import { MetalRingInformation } from './euring-codes/metal-ring-information-entity';
+import { EURINGOtherMarksInformation } from './euring-codes/euring-other-marks-information-entity';
+import { Species } from './euring-codes/species-entity';
+import { Manipulated } from './euring-codes/manipulated-entity';
+import { MovedBeforeTheCapture } from './euring-codes/moved-before-capture-entity';
+import { CatchingMethod } from './euring-codes/catching-method-entity';
+import { CathingLures } from './euring-codes/catching-lures-entity';
+import { Status } from './euring-codes/status-entity';
+import { Broodsize } from './euring-codes/broodsize-entity';
+import { PullusAge } from './euring-codes/pullus-age-entity';
+import { AccuracyOfPullusAge } from './euring-codes/accuracy-of-pullus-age-entity';
+import { PlaceCode } from './euring-codes/place-code-entity';
+import { Condition } from './euring-codes/condition-entity';
+import { Circumstances } from './euring-codes/circumstances-entity';
+import { CircumstancesPresumed } from './euring-codes/circumstances-presumed-entity';
+import { AccuracyOfDate } from './euring-codes/accuracy-of-date-entity';
+import { EURINGCodeIdentifier } from './euring-codes/euring-code-identifier-entity';
+import { ColorRingInformation } from './euring-codes/color-ring-information-entity';
 import { User } from './user-entity';
-import { StatusOfRing } from './status-of-ring-entity';
+import { StatusOfRing } from './euring-codes/status-of-ring-entity';
 import { Observation } from './observation-entity';
 
 @Entity()
